Package Creation for Salesforce Service Cloud (#4936)
* Package Creation for Salesforce Service Cloud

* Modified the UI text changes

* Removed -Preview from SolutionId
