+++ Solutions/Salesforce Service Cloud/Data Connectors/SalesforceSentinelConnector/function.json
