+++ Solutions/Salesforce Service Cloud/Data Connectors/SalesforceServiceCloud_API_FunctionApp.json
@@ -1,6 +1,6 @@
 {
     "id": "SalesforceServiceCloud",
-    "title": "Salesforce Service Cloud",
+    "title": "Salesforce Service Cloud (using Azure Function)",
     "publisher": "Salesforce",
     "descriptionMarkdown": "The Salesforce Service Cloud data connector provides the capability to ingest information about your Salesforce operational events into Azure Sentinel through the REST API. The connector provides ability to review events in your org on an accelerated basis, get [event log files](https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/event_log_file_hourly_overview.htm) in hourly increments for recent activity.",
     "additionalRequirementBanner":"These queries and workbooks are dependent on a parser based on Kusto to work as expected. Follow the steps to use this Kusto functions alias **SalesforceServiceCloud** in queries and workbooks. [Follow steps to get this Kusto functions>](https://aka.ms/sentinel-SalesforceServiceCloud-parser) ",
@@ -33,7 +33,7 @@
     ],
     "availability": {
         "status": 1,
-        "isPreview": true
+        "isPreview": false
     },
     "permissions": {
         "resourceProvider": [
