+++ Solutions/Salesforce Service Cloud/SolutionMetadata.json
@@ -1,0 +1,20 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-salesforceservicecloud",
+	"firstPublishDate": "2022-05-16",
+	"providers": ["Salesforce"],
+	"categories": {
+		"domains" : ["Cloud Provider"],
+		"verticals": []
+	},
+	"support": {
+	  "name": "Microsoft Corporation",
+	  "email": "support@microsoft.com",
+	  "tier": "Microsoft",
+	  "link": "https://support.microsoft.com/"
+	}
+}
+
+
+
+
